refactor(assistant-stream): keep ./ai-sdk to frontendTools, add ./internal
`wrapModelContentEnvelope`, `unwrapModelContentEnvelope`,
`ModelContentEnvelope`, `toAISDKContent` and `toAISDKDefaultOutput`
were never in the `@assistant-ui/ai-sdk` barrel; they crossed the
package boundary only because `@assistant-ui/ai-sdk` imports them.
Publishing them on `./ai-sdk` would make them permanent under the
append-only rule, and `toAISDKContent`'s return type is pinned to the
current provider spec. Move them to `assistant-stream/internal`
(the `@assistant-ui/core/internal` precedent) and leave `./ai-sdk`
at `frontendTools` + `FrontendTools`, which is what #6650 asked for.

The entry test now also pins that `./internal` never loads `ai`.

Review round 2 (github-actions review).
