+++ packages/ai-sdk/src/converters/convertMessage.ts
@@ -28,7 +28,7 @@ import {
   parsePartialJsonObject,
   type ReadonlyJSONObject,
 } from "assistant-stream/utils";
-import { unwrapModelContentEnvelope } from "assistant-stream/ai-sdk";
+import { unwrapModelContentEnvelope } from "assistant-stream/internal";
 
 type MessageMetadata = ThreadMessageLike["metadata"];
