+++ packages/ai-sdk/src/runtime/useAISDKRuntime.ts
@@ -52,7 +52,7 @@ import { toCreateMessage } from "../converters/toCreateMessage";
 import { vercelAttachmentAdapter } from "../adapters/vercelAttachmentAdapter";
 import { getVercelAIMessages } from "../utils/getVercelAIMessages";
 import { AISDKMessageConverter } from "../converters/convertMessage";
-import { wrapModelContentEnvelope } from "assistant-stream/ai-sdk";
+import { wrapModelContentEnvelope } from "assistant-stream/internal";
 import {
   type AISDKStorageFormat,
   aiSDKV6FormatAdapter,
