+++ packages/assistant-stream/src/ai-sdk/index.ts
@@ -1,7 +1,1 @@
 export { frontendTools, type FrontendTools } from "./frontendTools";
-export {
-  unwrapModelContentEnvelope,
-  wrapModelContentEnvelope,
-  type ModelContentEnvelope,
-} from "./modelContentEnvelope";
-export { toAISDKContent, toAISDKDefaultOutput } from "./toolOutputConversion";
